+++ src/actions/index.js
@@ -1,3 +1,4 @@
+export const SET_LAYERS = "SET_LAYERS";
 export const ADD_OR_UPDATE_LAYER = "ADD_LAYER";
 export const REMOVE_LAYER = "REMOVE_LAYER";
 
@@ -6,7 +7,12 @@ export const addOrUpdateLayer = ({layer}) => ({
     layer
 });
 
-export const removeLayer = ({ ukey }) => ({
+export const removeLayer = ({ukey}) => ({
     type: REMOVE_LAYER,
     ukey
+});
+
+export const setLayers = ({layers}) => ({
+    type: SET_LAYERS,
+    layers
 });