+++ src/components/SideBar.js
@@ -4,19 +4,7 @@ import LayersList from './layers/LayersList';
 
 class SideBar extends Component {
 
-	constructor(props) {
-		super(props);
-
-		this.forceUpdateLayers = this.forceUpdateLayers.bind(this);
-
-	};
-
-
-	forceUpdateLayers(layers) {
-		this.setState({layers: layers});
-		// this.refs.layerlist.forceUpdateLayers(layers);
-	};
-
+	// TODO: move menu bar here or consider removing this component
 
 	render() {
 		return (