+++ src/components/Map.js
@@ -1,10 +1,15 @@
+/* Main map component based on leaflet map.
+ *
+ *
+ */
 import React, {Component} from 'react'
+import {connect} from 'react-redux';
 import L from 'leaflet';
 import 'leaflet.heat';
 import 'leaflet.markercluster';
-
 import 'leaflet.markercluster/dist/MarkerCluster.css';
 import 'leaflet.markercluster/dist/MarkerCluster.Default.css';
+
 
 class Map extends Component {
 
@@ -15,55 +20,58 @@ class Map extends Component {
 			center: [49.8419, 24.0315],
 			zoom: 4,
 			layers: [
-			L.tileLayer('http://{s}.tile.osm.org/{z}/{x}/{y}.png', {
-				attribution: '&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
-			}),
+				L.tileLayer('http://{s}.tile.osm.org/{z}/{x}/{y}.png', {
+					attribution: '&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
+				}),
 			]
 		});
+
 		this.leafletMarkerLayers = {};
 		this.leafletPolylineLayers = {};
 		this.leafletHeatmapLayers = {};
 		this.leafletClusterLayers = {};
 	}
+
+
 	componentDidUpdate() {
-		let layers = Object.entries(this.props.layers);
+		let layers = this.props.layers;
 		let globalBounds = new L.LatLngBounds();
 		let ukeyMarkerArray = [];
 		let ukeyPolylineArray = [];
 		let ukeyHeatmapArray = [];
 		let ukeyClusterArray = [];
 		// Iterate through layers
-		layers.map(([, layer]) => {
-			if (layer.ukey !== undefined) {
-				globalBounds.extend(layer.bounds);
-				if (layer.rendering === "markers") {
-					ukeyMarkerArray.push(layer.ukey);
-					if (!this.leafletMarkerLayers[layer.ukey]) {
-						this.leafletMarkerLayers[layer.ukey] = L.layerGroup().addTo(this.map);
-					}
-					this.updateMarkerLayer(layer.data, layer.color, layer.ukey);
-				} else if (layer.rendering === "polyline") {
-					ukeyPolylineArray.push(layer.ukey);
-					if (this.leafletPolylineLayers[layer.ukey]) {
-						// todo find a way of updating the polyline layer instead of delete & recreate
-						this.map.removeLayer(this.leafletPolylineLayers[layer.ukey]);
-					}
-					this.updatePolylineLayer(layer.data, layer.color, layer.ukey);
-				} else if (layer.rendering === "heatmap") {
-					ukeyHeatmapArray.push(layer.ukey);
-					if (this.leafletHeatmapLayers[layer.ukey]) {
-						// todo find a way of updating the heat layer instead of delete & recreate
-						this.map.removeLayer(this.leafletHeatmapLayers[layer.ukey]);
-					}
-					this.updateHeatmapLayer(layer.data, layer.radius, layer.ukey)
-				} else if (layer.rendering === "clusters") {
-					ukeyClusterArray.push(layer.ukey);
-					if (!this.leafletClusterLayers[layer.ukey]) {
-						this.leafletClusterLayers[layer.ukey] = L.markerClusterGroup();
-						this.map.addLayer(this.leafletClusterLayers[layer.ukey]);
-					}
-					this.updateClusterLayer(layer.data, layer.color, layer.ukey);
+		layers.map((layer) => {
+			if (layer.ukey === undefined)
+				return null;
+			globalBounds.extend(layer.bounds);
+			if (layer.rendering === "markers") {
+				ukeyMarkerArray.push(layer.ukey);
+				if (!this.leafletMarkerLayers[layer.ukey]) {
+					this.leafletMarkerLayers[layer.ukey] = L.layerGroup().addTo(this.map);
 				}
+				this.updateMarkerLayer(layer.data, layer.color, layer.ukey);
+			} else if (layer.rendering === "polyline") {
+				ukeyPolylineArray.push(layer.ukey);
+				if (this.leafletPolylineLayers[layer.ukey]) {
+					// todo find a way of updating the polyline layer instead of delete & recreate
+					this.map.removeLayer(this.leafletPolylineLayers[layer.ukey]);
+				}
+				this.updatePolylineLayer(layer.data, layer.color, layer.ukey);
+			} else if (layer.rendering === "heatmap") {
+				ukeyHeatmapArray.push(layer.ukey);
+				if (this.leafletHeatmapLayers[layer.ukey]) {
+					// todo find a way of updating the heat layer instead of delete & recreate
+					this.map.removeLayer(this.leafletHeatmapLayers[layer.ukey]);
+				}
+				this.updateHeatmapLayer(layer.data, layer.radius, layer.ukey);
+			} else if (layer.rendering === "clusters") {
+				ukeyClusterArray.push(layer.ukey);
+				if (!this.leafletClusterLayers[layer.ukey]) {
+					this.leafletClusterLayers[layer.ukey] = L.markerClusterGroup();
+					this.map.addLayer(this.leafletClusterLayers[layer.ukey]);
+				}
+				this.updateClusterLayer(layer.data, layer.color, layer.ukey);
 			}
 			return null;
 		});
@@ -105,15 +113,35 @@ class Map extends Component {
 		});
 		this.map.flyToBounds(globalBounds);
 	}
+
+
+	addMarkerLayer(layer) {
+		if (!this.leafletMarkerLayers[layer.ukey]) {
+			this.leafletMarkerLayers[layer.ukey] = L.layerGroup().addTo(this.map);
+		}
+		this.updateMarkerLayer(layer.data, layer.color, layer.ukey);
+	}
+
+
+	deleteMarkerLayer(key) {
+		if (this.leafletMarkerLayers.includes(key)) {
+			this.map.removeLayer(this.leafletMarkerLayers[key]);
+			delete this.leafletMarkerLayers[key];
+			return true;
+		}
+		return false;
+	}
+
+
 	updateMarkerLayer(data, color, ukey) {
-		// todo check if the layer as change before rerendering it
+		// todo check if the layer has changed before rerendering it
 		this.leafletMarkerLayers[ukey].clearLayers();
 		let m = null;
 		let rgbColor = `rgb(${color.r}, ${color.g}, ${color.b})`;
 		data.forEach(entry => {
 			m = L.circleMarker(
 				entry.pos,
-				{ 
+				{
 					title: entry.tooltip,
 					fill: true,
 					radius: 5,
@@ -123,58 +151,77 @@ class Map extends Component {
 					fillOpacity: color.a
 				}
 			).addTo(this.leafletMarkerLayers[ukey]);
-			m.bindTooltip(entry.tooltip);
+			if (entry.tooltip !== undefined)
+				m.bindPopup(entry.tooltip);
 		});
 	}
+
+
 	updatePolylineLayer(data, color, ukey) {
-		// todo check if the layer as change before rerendering it
+		// todo check if the layer has changed before rerendering it
 		// this.leafletLayers[ukey].clearLayers();
 		let rgbColor = `rgb(${color.r}, ${color.g}, ${color.b})`;
-		let polylineData = [];
-		polylineData = data.map((entry) => {
+		let polylineData = data.map((entry) => {
 			return entry.pos;
 		});
-		this.leafletPolylineLayers[ukey] = L.polyline(polylineData, { color: rgbColor }).addTo(this.map);
+		var mapLayer = L.polyline(polylineData, {color: rgbColor}).addTo(this.map);
+		this.leafletPolylineLayers[ukey] = mapLayer;
 		// this.leafletPolylineLayers[ukey].setLatLngs(polylineData);
 		// this.leafletPolylineLayers[ukey].setConfig??({ color });
 	}
+
+
 	updateHeatmapLayer(data, radius,  ukey) {
-		// todo check if the layer as change before rerendering it
-		let heatData = [];
-		heatData = data.map((entry) => {
+		// todo check if the layer has changed before rerendering it
+		let heatData = data.map((entry) => {
 			return entry.pos.concat(1.0);
 		});
-		this.leafletHeatmapLayers[ukey] = L.heatLayer(heatData, {
+		var mapLayer = L.heatLayer(heatData, {
 			radius: radius,
-			 minOpacity: 0.1,
-			  blur: 15,
-			   max: 10.0
-			}).addTo(this.map);
+			minOpacity: 0.1,
+			blur: 15,
+			max: 10.0
+		}).addTo(this.map);
+		this.leafletHeatmapLayers[ukey] = mapLayer;
 		// this.leafletHeatmapLayers[ukey].setLatLngs(heatData);
 		// this.leafletHeatmapLayers[ukey].setConfig({ radius });
 	}
+
+
 	updateClusterLayer(data, color, ukey) {
-		// todo check if the layer as change before rerendering it
+		// todo check if the layer has changed before rerendering it
 		this.leafletClusterLayers[ukey].clearLayers();
 		let m = null;
 		let rgbColor = `rgb(${color.r}, ${color.g}, ${color.b})`;
 		data.forEach(entry => {
 			m = L.circleMarker(entry.pos,{
-				 title: entry.tooltip,
-				 fill: true,
-				 radius: 5,
-				 color: rgbColor,
-				 fillColor: rgbColor,
-				 opacity: color.a,
-				 fillOpacity: color.a
+				title: entry.tooltip,
+				fill: true,
+				radius: 5,
+				color: rgbColor,
+				fillColor: rgbColor,
+				opacity: color.a,
+				fillOpacity: color.a
 			});
-			m.bindTooltip(entry.tooltip);
+			if (entry.tooltip !== undefined)
+				m.bindPopup(entry.tooltip);
 			this.leafletClusterLayers[ukey].addLayer(m);
 		});
 	}
-	render() {	
+
+
+	render() {
 		return <div id="map"></div>;
 	}
 }
 
-export default  Map;
+
+const mapStateToProps = (state, ownProps) => {
+	return {
+		layers: state.layers,
+		...ownProps
+	}
+};
+
+
+export default connect(mapStateToProps)(Map);