+++ src/components/layers/LayerList.js
@@ -1,6 +1,7 @@
 import React, {Component} from 'react';
 import Accordion from 'react-bootstrap/Accordion';
-import Layer from './layer';
+import Layer from './Layer';
+import {connect} from 'react-redux';
 
 
 class LayersList extends Component {
@@ -9,12 +10,10 @@ class LayersList extends Component {
 		super(props);
 
 		this.state = {
-			driver: props.driver,
-			layers: props.layers
+			driver: props.driver
 		};
 
 		this.forceUpdateLayers = this.forceUpdateLayers.bind(this);
-		this.sendData = this.sendData.bind(this);
 		this.renderLayers = this.renderLayers.bind(this);
 		this.renderNewLayer = this.renderNewLayer.bind(this);
 		this.deleteLayer = this.deleteLayer.bind(this);
@@ -25,21 +24,6 @@ class LayersList extends Component {
 		this.setState({layers: layers});
 	};
 
-
-	sendData(data) {
-		/*Receives new data from child layer
-           and propagate it to parent
-        */
-		let new_layer = data.layer;
-		let layers = this.state.layers;
-		layers[new_layer.ukey] = new_layer;
-		this.setState({
-			layers: layers
-		});
-		this.props.sendData({
-			layers: layers
-		});
-	};
 
 
 	deleteLayer(ukey) {
@@ -59,8 +43,8 @@ class LayersList extends Component {
 
 
 	renderLayers() {
-		let layers = Object.entries(this.state.layers);
-		return layers.map(([, layer]) => {
+		// let layers = Object.entries(this.state.layers);
+		return this.props.layers.map((layer) => {
 			return (
 				<Layer data-id="layers" key={layer.ukey} ukey={layer.ukey} layer={layer} deleteLayer={this.deleteLayer}
 					   sendData={this.sendData} driver={this.state.driver}/>
@@ -88,5 +72,11 @@ class LayersList extends Component {
 	};
 }
 
+const mapStateToProps = (state, ownProps) => {
+	return {
+		layers: state.layers,
+		...ownProps
+	}
+};
 
-export default LayersList;
+export default connect(mapStateToProps)(LayersList);