+++ src/store/index.js
@@ -1,0 +1,12 @@
+import {combineReducers, createStore} from 'redux';
+import layerReducer from "../reducers/layers";
+// import driverReducer from "../reducers/driver";
+
+export default () => {
+    return createStore(
+        combineReducers({
+            layers: layerReducer,
+            // driver: driverReducer,
+        })
+    );
+};