+++ release-notes.md
@@ -4,6 +4,9 @@
 ## 0.5.0 (WIP)
 
 - Support for clusters in map rendering (#50)
+- Better support for Neo4j Desktop versions (#53)
+- Some code refactoring (#52)
+
 
 ## 0.4.0 (2020-02-27)